convert to es module and revise / refactor
